Move periodicCache to Utils package
