+++ freqtrade/util/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa: F401
+from freqtrade.util.periodic_cache import PeriodicCache
