+++ freqtrade/util/periodic_cache.py
